Module authzen_diesel_core::prelude

Re-exports

pub use crate::connection::TxCleanup;

pub use crate::connection::TxCleanupError;

pub use crate::connection::TxCleanupFn;

pub use crate::connection::TxFn;

pub use crate::deletable::*;

pub use crate::is_deleted::*;

pub use crate::audit::*;

pub use crate::deletable::*;

pub use crate::macros::*;

pub use crate::paginate::*;

pub use crate::schema::*;

pub use crate::pool::*;

bb8 or deadpool or mobc

Enums

DbEntityError

Wrapper error type returned from DbEntity methods solely exists to facilitate error conversion when converting between the raw db representation and the entity representation.

Traits

DbDelete

DbEntity

Base trait which unites common associated types used by the common sql operations in diesel (i.e. select, insert, update and delete).

DbGet

DbInsert

DbUpdate